# src/lib/openapi.yaml
openapi: 3.0.0
info:
  title: PetStayTion API
  version: 1.0.0
  description: API for pet care service management
servers:
  - url: /
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/advertisements:
    get:
      summary: Get all advertisements for the authenticated user
      description: |
        Returns all advertisements for the authenticated user associated with any of their service providers.
        Includes advertisements of all statuses. Only returns title, startDate, endDate, serviceStartTime, serviceEndTime, keyImage (first image), species, and the service provider's city information for each advertisement.
        Requires a valid Firebase authentication token.
      tags: [Advertisements]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Number of items per page
      responses:
        "200":
          description: Advertisements retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  advertisements:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        title:
                          type: string
                          example: "Profesjonalne wyprowadzanie psów w centrum Warszawy"
                        startDate:
                          type: string
                          format: date-time
                          example: "2025-09-26T00:00:00Z"
                        endDate:
                          type: string
                          format: date-time
                          nullable: true
                          example: "2025-11-25T00:00:00Z"
                        serviceStartTime:
                          type: string
                          nullable: true
                          example: "09:00"
                        serviceEndTime:
                          type: string
                          nullable: true
                          example: "17:00"
                        keyImage:
                          type: string
                          nullable: true
                          example: "https://images.unsplash.com/photo-1552053831-71594a27632d?w=500"
                        species:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: integer
                                example: 1
                              name:
                                type: string
                                example: "Pies"
                        city:
                          type: object
                          properties:
                            idCity:
                              type: integer
                              example: 1
                            name:
                              type: string
                              example: "Warszawa"
                            imageUrl:
                              type: string
                              nullable: true
                              example: "https://example.com/city.jpg"
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 20
                      total:
                        type: integer
                        example: 50
                      totalPages:
                        type: integer
                        example: 3
        "401":
          description: Unauthorized (invalid or missing token)
        "404":
          description: User not found
        "500":
          description: Internal server error
    post:
      summary: Create a new advertisement
      description: |
        Creates a new advertisement for the authenticated user who is an active service provider.
        Requires a valid Firebase authentication token.
        At least one image must be provided as multipart/form-data, and price must be a non-negative number if provided.
        speciesIds is optional array of species IDs.
      tags: [Advertisements]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - startDate
                - endDate
                - serviceId
                - images
              properties:
                title:
                  type: string
                  example: "Profesjonalne wyprowadzanie psów w centrum Warszawy"
                description:
                  type: string
                  nullable: true
                  example: "Oferuję profesjonalne wyprowadzanie psów..."
                price:
                  type: string
                  nullable: true
                  example: "25.0"
                startDate:
                  type: string
                  format: date-time
                  example: "2025-09-26T00:00:00Z"
                endDate:
                  type: string
                  format: date-time
                  example: "2025-11-25T00:00:00Z"
                serviceStartTime:
                  type: string
                  nullable: true
                  example: "1970-01-01T09:00:00Z"
                serviceEndTime:
                  type: string
                  nullable: true
                  example: "1970-01-01T17:00:00Z"
                serviceId:
                  type: integer
                  example: 1
                images:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    format: binary
                speciesIds:
                  type: string
                  description: JSON array of species IDs
                  example: "[1, 2]"
                  nullable: true
      responses:
        "200":
          description: Advertisement created successfully
        "400":
          description: Missing required fields, no images provided, or invalid price or speciesIds
        "401":
          description: Unauthorized (invalid or missing token)
        "403":
          description: Forbidden (user is not an active service provider)
        "404":
          description: User or service not found
        "409":
          description: Conflict (advertisement with this title already exists for the user)
        "500":
          description: Internal server error
...